Split up src/lib/embed.ts
